labels: add shared shard_flag_enabled helper
Move the per-task shard feature-flag reader out of the runtime crate's
task_schema module and into the shared labels crate, so both the V1 (runtime)
and V2 (runtime-next) task builders can read a flag from a task's
shard_template labels through one implementation. No behavior change; the V1
materialize and derive read-relaxation sites now call labels::shard_flag_enabled.
