+++ crates/runtime/src/derive/task.rs
@@ -1,5 +1,5 @@
 use super::{Task, Transform};
-use crate::task_schema::{relax_inferred_datetime_formats, shard_flag_enabled};
+use crate::task_schema::relax_inferred_datetime_formats;
 use anyhow::Context;
 use proto_flow::derive::{Request, Response, request, response};
 use proto_flow::flow;
@@ -50,7 +50,7 @@ impl Task {
         // Opt-in, per-task relaxation of read-side date-time `format`
         // enforcement inherited from each source collection's inferred schema.
         // See Transform::new and estuary/flow#3133.
-        let relax_inferred_datetime = shard_flag_enabled(
+        let relax_inferred_datetime = labels::shard_flag_enabled(
             shard_template.as_ref(),
             labels::RELAX_INFERRED_DATETIME_FLAG,
         );
