+++ crates/runtime/src/task_schema.rs
@@ -1,21 +1,6 @@
-// Helpers shared by the materialize and derive task builders for read-side
-// schema handling. See estuary/flow#3133.
-
-/// Return whether the shard feature-flag `flag` (a fully-qualified flag label
-/// such as [`labels::RELAX_INFERRED_DATETIME_FLAG`]) is set to "true" within a
-/// task's `shard_template` labels. This mirrors the Go-side `LabelSet.ValueOf`
-/// check used by the v2 runtime dispatch, but reads the flag at the Rust
-/// task-build layer where read-side validators are constructed.
-pub fn shard_flag_enabled(
-    shard_template: Option<&proto_gazette::consumer::ShardSpec>,
-    flag: &str,
-) -> bool {
-    shard_template
-        .and_then(|shard| shard.labels.as_ref())
-        .and_then(|set| labels::maybe_one(set, flag).ok())
-        .map(|value| value == "true")
-        .unwrap_or(false)
-}
+// Helper shared by the materialize and derive task builders for read-side
+// schema handling. See estuary/flow#3133. Per-task flag reading lives in
+// `labels::shard_flag_enabled`.
 
 /// Return `read_schema_json` with `date`/`date-time`/`time` `format` keywords
 /// stripped from its inlined inferred schema, leaving the rest of the bundle
@@ -33,32 +18,6 @@ pub fn relax_inferred_datetime_formats(
 #[cfg(test)]
 mod test {
     use super::*;
-
-    fn shard_with_flag(value: Option<&str>) -> proto_gazette::consumer::ShardSpec {
-        let labels =
-            value.map(|value| labels::build_set([(labels::RELAX_INFERRED_DATETIME_FLAG, value)]));
-        proto_gazette::consumer::ShardSpec {
-            labels,
-            ..Default::default()
-        }
-    }
-
-    #[test]
-    fn test_shard_flag_enabled() {
-        let flag = labels::RELAX_INFERRED_DATETIME_FLAG;
-
-        let on = shard_with_flag(Some("true"));
-        assert!(shard_flag_enabled(Some(&on), flag));
-
-        // A value other than "true" does not enable the flag.
-        let off = shard_with_flag(Some("false"));
-        assert!(!shard_flag_enabled(Some(&off), flag));
-
-        // Flag absent, no labels, or no shard template at all.
-        let none = shard_with_flag(None);
-        assert!(!shard_flag_enabled(Some(&none), flag));
-        assert!(!shard_flag_enabled(None, flag));
-    }
 
     // A read-schema bundle whose inlined inferred schema tags a field
     // `format: date-time`, shaped as the control plane assembles it.
